create ingredient database

+++ Shared/Models/UnitConversion.cs
@@ -12,8 +12,8 @@ namespace RecipEase.Shared.Models
         [ForeignKey("ConvertsFromUnit")]
         public string ConvertsFromUnitName { get; set; }
 
-        public int Ratio { get; set; }
-        
+        public double Ratio { get; set; }
+
 
         public Unit ConvertsToUnit { get; set; }
         public Unit ConvertsFromUnit { get; set; }
